Test behaviour of 64 byte keys passed to crypto_kdf_derive_from_key
Currently, sodium-native and sodium-javascript are returning different
hashes. The code in hyperdrive passes a 64 byte secret key to the kdf,
but only 32 bytes are used by the native version, but all 64 bytes are
used in the javascript version. As a result, hyperdrive secret keys
can't be imported/exported across the two sodium implementations.

https://gist.github.com/jimpick/3e869522eddaad77ac1bc9e64f36e1a7

This adds a fixture with a 64 byte key, which will pass on
sodium-native, but fail on sodium-javascript.