Add crypto_box_easy and crypto_box_open_easy
This includes a port of box_easy.c from libsodium, which gives us a few
small test-cases to ensure that our implementation works. I've made an
effort to document anywhere my tests diverge from the upstream tests,
and have tested this with both Sodium-Native and my development branch
of Sodium-JavaScript to ensure that these tests are passing.