Add tests for crypto_auth and crypto_auth_verify
The upstream libsodium tests used some methods that Sodium-Native
doesn't expose, like `crypto_auth_hmacsha512_init`, so the only thing
this shared with the libsodium tests is the input and expected output to
confirm that we have a byte-for-byte match. This file should probably be
more fleshed out in the future, but this should help confirm the
validity of future implementations in Sodium-JavaScript.